+++ ocrd/ocrd/workspace.py
@@ -9,6 +9,7 @@ from atomicwrites import atomic_write
 from deprecated.sphinx import deprecated
 
 from ocrd_models import OcrdMets, OcrdExif, OcrdFile
+from ocrd_models.ocrd_page import parse
 from ocrd_utils import (
     getLogger,
     image_from_polygon,
@@ -27,6 +28,7 @@ from ocrd_utils import (
     pushd_popd,
     MIME_TO_EXT,
     MIME_TO_PIL,
+    MIMETYPE_PAGE
 )
 
 from .workspace_backup import WorkspaceBackupManager
@@ -117,7 +119,7 @@ class Workspace():
             f.local_filename = f.url
             return f
 
-    def remove_file(self, ID, force=False, keep_file=False):
+    def remove_file(self, ID, force=False, keep_file=False, page_recursive=False, page_same_group=False):
         """
         Remove a file from the workspace.
 
@@ -125,12 +127,23 @@ class Workspace():
             ID (string|OcrdFile): ID of the file to delete or the file itself
             force (boolean): Continue removing even if file not found in METS
             keep_file (boolean): Whether to keep files on disk
+            page_recursive (boolean): Whether to remove all images referenced in the file if the file is a PAGE-XML document.
+            page_same_group (boolean): Remove only images in the same file group as the PAGE-XML. Has no effect unless ``page_recursive`` is ``True``.
         """
         log.debug('Deleting mets:file %s', ID)
         if isinstance(ID, OcrdFile):
             ID = ID.ID
         try:
             ocrd_file = self.mets.remove_file(ID)
+            if ocrd_file.mimetype == MIMETYPE_PAGE and page_recursive:
+                with pushd_popd(self.directory):
+                    ocrd_page = parse(self.download_file(ocrd_file).local_filename, silence=True)
+                    for img_url in ocrd_page.get_AllImagePaths():
+                        img_kwargs = {'url': img_url}
+                        if page_same_group:
+                            img_kwargs['fileGrp'] = ocrd_file.fileGrp
+                        for img_file in self.mets.find_files(**img_kwargs):
+                            self.remove_file(img_file, keep_file=keep_file, force=force)
             if not keep_file:
                 if not ocrd_file.local_filename:
                     log.warning("File not locally available %s", ocrd_file)
@@ -145,7 +158,7 @@ class Workspace():
             if not force:
                 raise e
 
-    def remove_file_group(self, USE, recursive=False, force=False, keep_files=False):
+    def remove_file_group(self, USE, recursive=False, force=False, keep_files=False, page_recursive=False, page_same_group=False):
         """
         Remove a fileGrp.
 
@@ -154,12 +167,14 @@ class Workspace():
             recursive (boolean): Whether to recursively delete all files in the group
             force (boolean): Continue removing even if group or containing files not found in METS
             keep_files (boolean): When deleting recursively whether to keep files on disk
+            page_recursive (boolean): Whether to remove all images referenced in the file if the file is a PAGE-XML document.
+            page_same_group (boolean): Remove only images in the same file group as the PAGE-XML. Has no effect unless ``page_recursive`` is ``True``.
         """
         if USE not in self.mets.file_groups and not force:
             raise Exception("No such fileGrp: %s" % USE)
         if recursive:
             for f in self.mets.find_files(fileGrp=USE):
-                self.remove_file(f.ID, force=force, keep_file=keep_files)
+                self.remove_file(f, force=force, keep_file=keep_files, page_recursive=page_recursive, page_same_group=page_same_group)
         if USE in self.mets.file_groups:
             self.mets.remove_file_group(USE)
         # XXX this only removes directories in the workspace if they are empty